% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_export_filename}
\alias{build_export_filename}
\title{Construit un nom de fichier standardisé pour les exports}
\usage{
build_export_filename(objet, suffixe, ext = "xlsx")
}
\arguments{
\item{objet}{Nom du contenu exporté (ex: \code{"masselongueur"}, \code{"cpue_tous"})}

\item{suffixe}{Résultat de \code{generate_filename_suffix()}}

\item{ext}{Extension souhaitée (ex: \code{"xlsx"}, \code{"png"})}
}
\value{
Une chaîne de caractères représentant le nom de fichier.
}
\description{
Exemple : \code{"masselongueur_PENT_2022_LacArchambault_no01565.xlsx"}
}
