% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_compare_capture_specimen.R
\name{cpue_compare_capture_specimen}
\alias{cpue_compare_capture_specimen}
\title{Comparer le nombre de captures et de spécimens par station}
\usage{
cpue_compare_capture_specimen(capture, specimen)
}
\arguments{
\item{capture}{Un \code{data.frame} contenant au moins les colonnes \code{no_station}
et \code{nb_capture}.}

\item{specimen}{Un \code{data.frame} contenant au moins la colonne \code{no_station}.}
}
\value{
Une liste contenant \code{success}, \code{message}, \code{data} et \code{flextable}.
}
\description{
Cette fonction compare, pour chaque station, le nombre de captures indiqué
dans la table \code{capture} et le nombre d'individus présents dans la table
\code{specimen}.
}
