% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render.R
\name{download_button_ui}
\alias{download_button_ui}
\title{Crée un bouton de téléchargement simple}
\usage{
download_button_ui(id, label = "Télécharger (.xlsx)")
}
\arguments{
\item{id}{Identifiant du bouton (doit correspondre à celui utilisé dans \verb{render_download_*})}

\item{label}{Texte affiché (défaut : "Télécharger (.xlsx)")}
}
\value{
Un bouton téléchargeable (à insérer dans \code{app_ui()})
}
\description{
Crée un bouton de téléchargement simple
}
