% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Enregistre un data.frame en fichier Excel (.xlsx)}
\usage{
download_data(data, path)
}
\arguments{
\item{data}{Un data.frame ou une liste de data.frames}

\item{path}{Chemin de sortie du fichier .xlsx}
}
\value{
NULL (fichier écrit sur disque)
}
\description{
Enregistre un data.frame en fichier Excel (.xlsx)
}
