% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{extract_from_env}
\alias{extract_from_env}
\title{Extraire la valeur de L∞ à partir de l’environnement du modèle}
\usage{
extract_from_env(mod)
}
\arguments{
\item{mod}{Un objet \code{nls} contenu dans \code{growth()}}
}
\value{
Un nombre correspondant à L∞
}
\description{
Fonction interne utilisée par \code{croissance_compare_modele()} pour lire \code{Sinf}
depuis l’environnement du modèle.
}
\keyword{internal}
