% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{extract_growth_convergence}
\alias{extract_growth_convergence}
\title{Extraire l'état de convergence d'un modèle de croissance}
\usage{
extract_growth_convergence(mod)
}
\arguments{
\item{mod}{Un objet de modèle (\code{nls}) provenant de \code{growth()}.}
}
\value{
Une chaîne de caractères décrivant l'état de convergence.
}
\description{
Extraire l'état de convergence d'un modèle de croissance
}
\keyword{internal}
