% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{extract_growth_model}
\alias{extract_growth_model}
\title{Extraire un modèle individuel depuis l'objet retourné par \code{growth()}}
\usage{
extract_growth_model(result, component)
}
\arguments{
\item{result}{Objet retourné par \code{growth()}.}

\item{component}{Nom de la composante à extraire (\code{"vout"}, \code{"gout"},
\code{"lout"}).}
}
\value{
Le modèle demandé, ou \code{NULL} si absent.
}
\description{
Extraire un modèle individuel depuis l'objet retourné par \code{growth()}
}
\keyword{internal}
