% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{extract_param}
\alias{extract_param}
\title{Extraire un paramètre et son intervalle de confiance à partir d’un modèle}
\usage{
extract_param(res, index)
}
\arguments{
\item{res}{Un objet de modèle (\code{nls}) provenant de \code{growth()}}

\item{index}{Position du paramètre (1 = L∞, 2 = K, 3 = t0)}
}
\value{
Un vecteur numérique ou NA en cas d’erreur
}
\description{
Fonction interne utilisée par \code{croissance_compare_modele()} pour obtenir un
intervalle de confiance pour un paramètre donné (1 = L∞, 2 = K, 3 = t0).
}
\keyword{internal}
