% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_num_fr}
\alias{format_num_fr}
\title{Formatter un nombre avec virgule décimale}
\usage{
format_num_fr(x, digits = 2, na = "-")
}
\arguments{
\item{x}{Un vecteur numérique.}

\item{digits}{Nombre de décimales à afficher.}

\item{na}{Valeur à afficher si \code{NA}.}
}
\value{
Un vecteur caractère formaté.
}
\description{
Convertit un nombre en texte avec un nombre fixe de décimales,
en utilisant la virgule comme séparateur décimal.
}
\examples{
format_num_fr(1.2)
format_num_fr(1.2, digits = 3)

}
