% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_filename_suffix}
\alias{generate_filename_suffix}
\title{Génère un suffixe de nom de fichier à partir des métadonnées du lac}
\usage{
generate_filename_suffix(typ_pech, annee, no_lac, nom_lac = NULL)
}
\arguments{
\item{typ_pech}{Code du type de pêche (ex: \code{"PENT"})}

\item{annee}{Année de l'inventaire (ex: \code{2022})}

\item{no_lac}{Numéro du lac (ex: \code{"01565"})}

\item{nom_lac}{Nom du lac (ex: \code{"Lac Archambault"}). Optionnel.}
}
\value{
Une chaîne de caractères.
}
\description{
Exemple : \code{"PENT_2022_LacArchambault_no01565"}
}
