% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{get_growth_start_values}
\alias{get_growth_start_values}
\title{Déterminer les valeurs initiales à utiliser pour les modèles de croissance}
\usage{
get_growth_start_values(df)
}
\arguments{
\item{df}{Un \code{data.frame} contenant au minimum les colonnes \code{ltm} et \code{age}.}
}
\value{
Une liste contenant :
\describe{
\item{pi}{Une liste nommée avec \code{Linf}, \code{K} et \code{t0}}
\item{message}{Un message d’avertissement à afficher à l’utilisateur, ou
\code{NULL} si \code{vbStarts()} a réussi}
}
}
\description{
Cette fonction tente d’abord d’estimer automatiquement les valeurs initiales
des paramètres de croissance (\code{Linf}, \code{K}, \code{t0}) à l’aide de \code{vbStarts()}.
}
\details{
Dans certains cas (ex. structure d’âge déséquilibrée ou peu informative),
cette estimation peut échouer. Une stratégie de rechange est alors utilisée
afin de permettre la modélisation :
\itemize{
\item \code{Linf} : longueur maximale observée dans l’échantillon
\item \code{K} : valeur fixée à 0.3
\item \code{t0} : valeur fixée à 0
}

Ces valeurs par défaut sont issues de recommandations générales (Ogle 2016)
et permettent d’initialiser les modèles même lorsque les données ne permettent
pas une estimation automatique robuste.
}
\keyword{internal}
