% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wri.R
\name{get_wr_constants}
\alias{get_wr_constants}
\title{Récupérer les constantes Wr pour une espèce donnée}
\usage{
get_wr_constants(espece)
}
\arguments{
\item{espece}{Code d'espèce, par exemple \code{"SANA"}, \code{"SAFO"} ou \code{"SAVI"}.}
}
\value{
Un \code{data.frame} d'une ligne contenant les constantes de calcul du Wr.
}
\description{
Cette fonction retourne les coefficients de référence pour le calcul de
l'indice de condition relatif (Wr) pour une espèce supportée.
}
\keyword{internal}
