% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_error}
\alias{handle_error}
\title{Gérer les erreurs silencieusement}
\usage{
handle_error(e)
}
\arguments{
\item{e}{Objet d’erreur (\code{condition}) attrapé par \code{tryCatch}.}
}
\value{
Message de l’erreur (chaîne de caractères).
}
\description{
Fonction utilitaire pour extraire le message d’erreur dans un \code{tryCatch}.
}
