% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{is_growth_model_available}
\alias{is_growth_model_available}
\title{Vérifier si un modèle de croissance retourné par \code{growth()} est disponible}
\usage{
is_growth_model_available(mod)
}
\arguments{
\item{mod}{Objet supposé correspondre à un modèle individuel (\code{nls}).}
}
\value{
\code{TRUE} si le modèle semble disponible, \code{FALSE} sinon.
}
\description{
Vérifier si un modèle de croissance retourné par \code{growth()} est disponible
}
\keyword{internal}
