% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelled_data}
\alias{labelled_data}
\title{Extraire les labels comme noms de colonnes (pour usage dans export)}
\usage{
labelled_data(data)
}
\arguments{
\item{data}{Un \code{data.frame} avec des labels (\code{labelled}).}
}
\value{
Un \code{data.frame} avec noms de colonnes remplacés par leurs labels.
}
\description{
Extraire les labels comme noms de colonnes (pour usage dans export)
}
