% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_eval_modele.R
\name{maturite_eval_modele}
\alias{maturite_eval_modele}
\title{Évaluer l'ajustement des modèles de maturité (L50 ou A50)}
\usage{
maturite_eval_modele(models)
}
\arguments{
\item{models}{Liste des modèles retournée}
}
\value{
Un \code{data.frame} avec les critères d’évaluation, trié par AICc
}
\description{
Évaluer l'ajustement des modèles de maturité (L50 ou A50)
}
