% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_prepare.R
\name{maturite_prepare}
\alias{maturite_prepare}
\title{Prépare les données pour l’ajustement d’un modèle de maturité (L50 ou A50)}
\usage{
maturite_prepare(specimen_data, variable = c("ltm", "age"))
}
\arguments{
\item{specimen_data}{Jeu de données contenant les colonnes \code{maturite}, \code{sexe}, et la variable quantitative (\code{ltm} ou \code{age})}

\item{variable}{Variable quantitative à utiliser : \code{"ltm"} (par défaut) ou \code{"age"}}
}
\value{
Un data.frame filtré et transformé prêt pour le modèle
}
\description{
Prépare les données pour l’ajustement d’un modèle de maturité (L50 ou A50)
}
