% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_get_age_max.R
\name{mortalite_get_age_max}
\alias{mortalite_get_age_max}
\title{Trouver l’âge maximal dans les données de mortalité}
\usage{
mortalite_get_age_max(data)
}
\arguments{
\item{data}{Un \code{data.frame} contenant les données de spécimens. Doit inclure une colonne nommée \code{age} de type numérique ou entier.}
}
\value{
Un entier (\code{integer}) correspondant à l’âge maximal observé. Retourne \code{NA_integer_} si aucun âge valide n’est présent.
}
\description{
Cette fonction retourne l’âge maximal observé dans les données de mortalité, en ignorant les valeurs manquantes.
Elle vérifie également la présence de la colonne \code{age} et la présence de données valides.
}
\examples{
# Exemple avec des âges valides
df <- data.frame(age = c(1, 2, 3, NA, 4))
mortalite_get_age_max(df)

# Exemple sans âges valides
df2 <- data.frame(age = c(NA, NA))
mortalite_get_age_max(df2)

# Exemple sans colonne 'age'
try(mortalite_get_age_max(data.frame(other = 1:5)))

}
