% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_get_peak_plus.R
\name{mortalite_get_peak_plus}
\alias{mortalite_get_peak_plus}
\title{Déterminer l’âge Peak Plus à partir de la structure d’âge}
\usage{
mortalite_get_peak_plus(data)
}
\arguments{
\item{data}{Un \code{data.frame} contenant une colonne \code{age} et déjà filtré pour une seule espèce.}
}
\value{
Un entier correspondant au Peak Plus
}
\description{
Cette fonction retourne l’âge le plus fréquent (le mode) dans la structure d’âge,
augmenté de 1 (selon la définition du Peak Plus).
}
