% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_fonctions_externes.R
\name{o.r.test}
\alias{o.r.test}
\title{Test d'ajustement de Osius & Rojek}
\usage{
o.r.test(obj)
}
\arguments{
\item{obj}{Un objet de modèle de type \code{glm} ajusté avec une famille binomiale.}
}
\value{
Une valeur p (\code{p.value}) du test d’overdispersion.
}
\description{
Test d'ajustement de Osius & Rojek
}
