% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{pen_constants}
\alias{pen_constants}
\title{Constantes associées aux espèces suivies par PEN}
\format{
Un \code{tibble} avec une ligne par espèce.
}
\usage{
pen_constants
}
\description{
Contient le nom commun, le binwidth recommandé pour les histogrammes,
les seuils de classes PSD (\code{breaks}) et leurs libellés (\code{break_labels}) par espèce (\code{sp}).
}
\keyword{datasets}
