% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{psd_classnames}
\alias{psd_classnames}
\title{Noms standardisés des classes PSD}
\format{
An object of class \code{character} of length 6.
}
\usage{
psd_classnames
}
\description{
Utilisés dans les fonctions PSD (\code{psd_q}, \code{psd_byclass}, etc.)
}
\keyword{datasets}
