% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render.R
\name{render_download_plot}
\alias{render_download_plot}
\title{Crée un bouton de téléchargement pour un graphique ggplot}
\usage{
render_download_plot(
  id,
  plot,
  filename = NULL,
  filename_suffix = NULL,
  width = 7,
  height = 5,
  dpi = 300,
  label = "Télécharger le graphique"
)
}
\arguments{
\item{id}{Identifiant du bouton}

\item{plot}{Objet ggplot ou \code{reactive()} le retournant}

\item{filename}{Nom complet du fichier PNG (sans extension)}

\item{filename_suffix}{Suffixe à ajouter automatiquement (ex: \code{"PENT_2022_LacX"})}

\item{width}{Largeur (en pouces)}

\item{height}{Hauteur (en pouces)}

\item{dpi}{Résolution (DPI)}

\item{label}{Libellé du bouton}
}
\description{
Crée un bouton de téléchargement pour un graphique ggplot
}
