% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render.R
\name{render_download_table}
\alias{render_download_table}
\title{Crée un bouton de téléchargement pour un tableau (.xlsx)}
\usage{
render_download_table(id, data, filename = NULL, label = NULL)
}
\arguments{
\item{id}{Identifiant du bouton (ex: "masselongueur_dl")}

\item{data}{Un objet \code{data.frame} ou un \code{reactive()} le retournant}

\item{filename}{Nom du fichier à enregistrer (.xlsx). Peut être une chaîne ou un \code{reactive()}}

\item{label}{(non utilisé ici – texte défini dans le \code{ui})}
}
\value{
Un \code{downloadHandler} affecté à l’objet \code{output[[id]]}
}
\description{
Cette fonction crée dynamiquement un bouton de téléchargement (Shiny) qui permet d’exporter un \code{data.frame}
en format \code{.xlsx}, à partir d’un objet réactif ou non. Le bouton doit être déclaré côté \code{ui} avec l’\code{id} fourni.
}
