% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render.R
\name{render_table_flextable}
\alias{render_table_flextable}
\title{Affiche un tableau flextable dans une application Shiny}
\usage{
render_table_flextable(output_id, flextable)
}
\arguments{
\item{output_id}{Identifiant utilisé dans \code{uiOutput()} pour insérer le tableau.}

\item{flextable}{Une expression, un objet \code{flextable}, ou une fonction \code{reactive()}.}
}
\description{
Cette fonction permet d’insérer dynamiquement un tableau \code{flextable}
dans une interface Shiny, à l’intérieur d’un \code{uiOutput()}
identifié par \code{output_id}.
}
