% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render.R
\name{render_user_guide_if_needed}
\alias{render_user_guide_if_needed}
\title{Rendre le fichier user_guide.rmd en HTML (si nécessaire)}
\usage{
render_user_guide_if_needed(
  rmd_path = "texte/user_guide.rmd",
  html_path = "www/user_guide.html"
)
}
\arguments{
\item{rmd_path}{Chemin vers le fichier .Rmd}

\item{html_path}{Chemin vers le fichier .html à générer}
}
\description{
Vérifie si le fichier HTML est plus vieux que le .Rmd, et le recompilé si besoin.
Utile pour afficher dans l'app Shiny via iframe.
}
