% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{sans_warning_proba}
\alias{sans_warning_proba}
\title{Exécute une expression glm() en filtrant le warning "probabilités ajustées à 0 ou 1"}
\usage{
sans_warning_proba(expr)
}
\arguments{
\item{expr}{Une expression \code{glm()} passée sans guillemets}
}
\value{
Le résultat de \code{glm()}, sans émettre de warning si celui-ci correspond à l'ajustement numérique
}
\description{
Exécute une expression glm() en filtrant le warning "probabilités ajustées à 0 ou 1"
}
\keyword{internal}
