% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_age.R
\name{structure_age}
\alias{structure_age}
\title{Structure d'âge des spécimens : graphique, tableau brut et tableau formaté}
\usage{
structure_age(data, groupement = "tous")
}
\arguments{
\item{data}{Un \code{data.frame} contenant les colonnes suivantes :
\itemize{
\item \code{sp} : code de l’espèce (doit être unique),
\item \code{age} : âge numérique du spécimen,
\item la variable de regroupement choisie (\code{sexe}, \code{marquage} ou \code{maturite}), si applicable.
}}

\item{groupement}{Type de regroupement à utiliser pour la coloration du graphique :
\code{"tous"} (par défaut), \code{"sexe"}, \code{"maturite"} ou \code{"marquage"}.}
}
\value{
Une liste nommée contenant trois éléments :
\describe{
\item{plot}{Objet \code{ggplot} représentant l’histogramme de la structure d’âge.}
\item{data}{\code{data.frame} avec les âges comptés (\code{age}, \code{n}).}
\item{flextable}{Tableau formaté avec \code{flextable}.}
}
}
\description{
Cette fonction produit la distribution d’âge d’une espèce sous forme de graphique (histogramme),
d’un tableau brut (\code{data.frame}) et d’un tableau formaté (\code{flextable}). Elle prend en charge un regroupement
facultatif par sexe, marquage ou maturité.
}
\examples{
df <- data.frame(sp = "SANA", age = c(1, 2, 2, 3, 3, 3))
structure_age(df, groupement = "tous")
}
