% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{wr_constants}
\alias{wr_constants}
\title{Constantes pour le calcul de l'indice de condition (Wr)}
\format{
Un \code{tibble} avec les colonnes :
\itemize{
\item \code{sp} : Code d’espèce (SANA, SAFO, SAVI)
\item \code{species} : Nom anglais de l'espèce
\item \code{min_TL} : Longueur minimale (mm)
\item \code{int} : Intercept de la régression log-log
\item \code{slope} : Pente de la régression log-log
\item \code{source} : Référence source
}
}
\usage{
wr_constants
}
\description{
Source : FSA::wsVal() pour Lake Trout, Brook Trout, Walleye
}
\keyword{datasets}
