% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpue_generate_biomasse.R
\name{bpue_generate_biomasse}
\alias{bpue_generate_biomasse}
\title{Créer une table de biomasse et BPUE par groupe biologique}
\usage{
bpue_generate_biomasse(data_specimen, data_station)
}
\arguments{
\item{data_specimen}{Un \code{data.frame} de spécimens filtrés (issu de \code{load_specimen()}).}

\item{data_station}{Un \code{data.frame} des stations valides (issu de \code{load_station()}).}
}
\value{
Une liste contenant :
\describe{
\item{\code{data}}{Un \code{data.frame} résumant la biomasse totale, la proportion, la BPUE et les intervalles de confiance pour chaque groupe biologique.}
\item{\code{flextable}}{Une version formatée du tableau pour l’exportation (Word, Shiny, etc.) à l’aide du package \code{flextable}.}
}
}
\description{
Cette fonction calcule la biomasse totale (kg), la proportion (\%) et la biomasse
par unité d'effort (BPUE, en kg/station) pour différents groupes biologiques d'une espèce cible.
}
