% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_export_filename}
\alias{build_export_filename}
\title{Construit un nom de fichier standardisé pour les exports}
\usage{
build_export_filename(objet, suffixe, ext = "xlsx")
}
\arguments{
\item{objet}{Nom du contenu exporté (ex: "masselongueur", "cpue_tous")}

\item{suffixe}{Résultat de \code{generate_filename_suffix()} ou \code{filename_suffix()}}

\item{ext}{Extension (xlsx, png, csv, etc.)}
}
\value{
Un nom complet comme "masselongueur_PENT_2022_LacArchambault_no01565.xlsx"
}
\description{
Construit un nom de fichier standardisé pour les exports
}
