% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_eval_modele.R
\name{build_individual_model_row}
\alias{build_individual_model_row}
\title{Évaluer les critères d’un modèle individuel L50}
\usage{
build_individual_model_row(mod, id)
}
\arguments{
\item{mod}{Un objet de type \code{glm} (ou \code{NULL} si le modèle n’a pas pu être ajusté).}

\item{id}{Identifiant du modèle (ex. combinaison lien + variable).}
}
\value{
Un \code{data.frame} avec les critères d’ajustement pour un seul modèle.
}
\description{
Fonction interne utilisée pour extraire les critères d’ajustement
(convergence, p-valeurs, AICc, etc.) d’un modèle de type \code{glm} utilisé
dans l’évaluation de la maturité (L50).
}
\keyword{internal}
