% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_abondance_table.R
\name{cpue_abondance_table}
\alias{cpue_abondance_table}
\title{Créer une table d’abondance structurée par groupe biologique}
\usage{
cpue_abondance_table(
  data,
  cpue_table_tous,
  cpue_table_femelles,
  best_model_tous,
  best_model_femelles
)
}
\arguments{
\item{data}{Un \code{data.frame} de spécimens filtrés pour le lac, année, etc.}

\item{cpue_table_tous}{Un \code{data.frame} issu de \code{cpue_compare_modele(...)$data} (tous spécimens)}

\item{cpue_table_femelles}{Idem pour les femelles matures}

\item{best_model_tous}{Nom du meilleur modèle (ex: "nb1") pour tous}

\item{best_model_femelles}{Nom du meilleur modèle (ex: "nb2") pour femelles}
}
\value{
Une liste avec deux éléments : \code{data} (tableau brut) et \code{flextable} (tableau formaté).
}
\description{
Cette fonction calcule les effectifs et proportions de différents groupes (tous, par sexe,
par statut reproducteur) à partir des spécimens d’une espèce cible. Elle ajoute les
colonnes de CPUE et d’intervalle de confiance (IC 95\%) en extrayant les données des meilleurs modèles.
}
