% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_abondance_table.R
\name{cpue_abondance_table}
\alias{cpue_abondance_table}
\title{Créer un tableau d’abondance structuré par groupe biologique}
\usage{
cpue_abondance_table(
  data,
  cpue_table_tous,
  cpue_table_femelles,
  best_model_tous,
  best_model_femelles
)
}
\arguments{
\item{data}{Un \code{data.frame} de spécimens filtrés, contenant minimalement les colonnes \code{sexe} et \code{maturite}.}

\item{cpue_table_tous}{Résultats des modèles CPUE pour tous les individus (format = "data.frame").}

\item{cpue_table_femelles}{Résultats des modèles CPUE pour les femelles reproductrices.}

\item{best_model_tous}{Nom du meilleur modèle CPUE pour le groupe "Tous".}

\item{best_model_femelles}{Nom du meilleur modèle CPUE pour le groupe "Repro. actifs femelles".}
}
\value{
Une liste contenant :
\describe{
\item{data}{Un \code{data.frame} résumant les abondances, proportions, CPUE et IC 95 \% par groupe.}
\item{flextable}{Un objet \code{flextable} prêt à l’exportation.}
}
}
\description{
Cette fonction calcule les effectifs et proportions de différents groupes biologiques
à partir d’un jeu de spécimens filtré (une espèce cible, un lac, une année, etc.).
Elle ajoute les colonnes de CPUE et d'intervalles de confiance (IC 95 \%) extraites des meilleurs modèles.
}
