% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_compare_modele.R
\name{cpue_compare_modele}
\alias{cpue_compare_modele}
\title{Comparer les modèles de CPUE et recommander le meilleur}
\usage{
cpue_compare_modele(cpue_data)
}
\arguments{
\item{cpue_data}{Un \code{data.frame} produit par \code{cpue_prepare()} contenant les colonnes \code{no_station} et \code{CPUE}.}
}
\value{
Une liste avec deux éléments : \code{data} (tableau brut) et \code{flextable} (tableau formaté).
}
\description{
Cette fonction ajuste cinq modèles (Poisson, NB1, NB2, CMP, GP) sur les données de CPUE
et retourne une liste contenant un tableau comparatif brut (\code{data.frame}) et sa version formatée (\code{flextable}).
}
