% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_fit_modele_cmp.R
\name{cpue_fit_modele_cmp}
\alias{cpue_fit_modele_cmp}
\title{Ajuster un modèle de CPUE de type CMP (Conway-Maxwell-Poisson)}
\usage{
cpue_fit_modele_cmp(cpue_data)
}
\arguments{
\item{cpue_data}{Un \code{data.frame} produit par \code{cpue_prepare()} contenant au minimum
les colonnes \code{no_station} et \code{cpue}.}
}
\value{
Un \code{data.frame} d’une seule ligne résumant le modèle ajusté, avec les colonnes suivantes :
\describe{
\item{methode}{Type de modèle utilisé (\code{"cmp"})}
\item{ajustement_hnp}{Pourcentage moyen d’observations hors bande du test HNP}
\item{aicc}{Critère d'information corrigé (aicc)}
\item{cpue_moyenne}{Valeur moyenne prédite par le modèle (exponentielle du lien)}
\item{ic_95}{Intervalle de confiance à 95 \% sous forme de chaîne de caractères}
\item{commentaire}{Texte interprétant l’ajustement : bon, marginal ou mauvais}
\item{convergence}{État de convergence du modèle (\code{TRUE} ou \code{FALSE})}
\item{nb_iterations_hnp}{Nombre total d’itérations HNP effectuées (2 ou 5)}
}
}
\description{
Cette fonction ajuste un modèle CMP via \code{glmmTMB} sur les données de CPUE par station.
Elle effectue également un test HNP pour évaluer la qualité de l’ajustement.
}
