% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_fit_modele_cmp.R
\name{cpue_fit_modele_cmp}
\alias{cpue_fit_modele_cmp}
\title{Ajuster un modèle de CPUE de type CMP (Conway-Maxwell-Poisson)}
\usage{
cpue_fit_modele_cmp(cpue_data)
}
\arguments{
\item{cpue_data}{Un \code{data.frame} produit par \code{cpue_prepare()} contenant au minimum :
\code{no_station}, \code{CPUE}.}
}
\value{
Un \code{data.frame} d’une ligne résumant le modèle ajusté.
}
\description{
Cette fonction ajuste un modèle CMP via \code{glmmTMB} sur les données de CPUE par station.
Elle effectue également un test HNP pour évaluer la qualité de l’ajustement.
}
