% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_fit_modele_gp.R
\name{cpue_fit_modele_gp}
\alias{cpue_fit_modele_gp}
\title{Ajuster un modèle de CPUE de type GP (Generalized Poisson)}
\usage{
cpue_fit_modele_gp(cpue_data)
}
\arguments{
\item{cpue_data}{Un \code{data.frame} produit par \code{cpue_prepare()} contenant au minimum :
\code{no_station}, \code{CPUE}.}
}
\value{
Un \code{data.frame} d’une ligne résumant le modèle ajusté.
}
\description{
Cette fonction ajuste un modèle glmmTMB avec distribution Generalized Poisson (GP)
sur les données de CPUE par station. Elle applique aussi un test HNP
(Half-Normal Plot) pour évaluer la qualité de l’ajustement.
}
