% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_fit_modele_nb1.R
\name{cpue_fit_modele_nb1}
\alias{cpue_fit_modele_nb1}
\title{Ajuster un modèle de CPUE de type NB1 (Negative Binomial 1)}
\usage{
cpue_fit_modele_nb1(cpue_data)
}
\arguments{
\item{cpue_data}{Un \code{data.frame} produit par \code{cpue_prepare()} contenant au minimum :
\code{no_station}, \code{CPUE}.}
}
\value{
Un \code{data.frame} d’une ligne résumant le modèle ajusté.
}
\description{
Cette fonction ajuste un modèle glmmTMB avec distribution NB1 sur les données de CPUE par station.
Elle réalise aussi un test HNP (Half-Normal Plot) pour vérifier l’ajustement du modèle.
}
