% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_fit_modele_nb2.R
\name{cpue_fit_modele_nb2}
\alias{cpue_fit_modele_nb2}
\title{Ajuster un modèle de CPUE de type NB2 (Negative Binomial 2)}
\usage{
cpue_fit_modele_nb2(cpue_data)
}
\arguments{
\item{cpue_data}{Un \code{data.frame} produit par \code{cpue_prepare()} contenant au minimum
les colonnes \code{no_station} (identifiant de la station) et \code{CPUE} (valeur numérique de la capture par unité d'effort).}
}
\value{
Un \code{data.frame} d’une ligne contenant :
\describe{
\item{methode}{Nom du modèle ("nb2")}
\item{ajustement_hnp}{Pourcentage moyen d'observations hors bande dans le HNP}
\item{aicc}{Critère d’information corrigé (AICc)}
\item{cpue_moyenne}{Estimation moyenne de la CPUE (valeur prédite)}
\item{ic_95}{Intervalle de confiance à 95 \%}
\item{commentaire}{Appréciation qualitative de l’ajustement}
\item{convergence}{Toujours TRUE si le modèle a été ajusté sans erreur}
\item{nb_iterations_hnp}{Nombre d’itérations du test HNP (2 ou 5)}
}
}
\description{
Cette fonction ajuste un modèle de régression NB2 (\code{glm.nb}) sur les données de CPUE par station,
puis applique un test HNP (Half-Normal Plot) pour évaluer la qualité de l’ajustement.
}
\examples{
fake_data <- tibble::tibble(
  no_station = 1:10,
  CPUE = rnbinom(10, mu = 5, size = 1)
)
cpue_fit_modele_nb2(fake_data)

}
