% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_fit_modele_poisson.R
\name{cpue_fit_modele_poisson}
\alias{cpue_fit_modele_poisson}
\title{Ajuster un modèle de CPUE de type Poisson}
\usage{
cpue_fit_modele_poisson(cpue_data)
}
\arguments{
\item{cpue_data}{Un \code{data.frame} produit par \code{cpue_prepare()} contenant au minimum :
\code{no_station}, \code{CPUE}.}
}
\value{
Un \code{data.frame} d’une ligne résumant le modèle ajusté, avec les colonnes :
\describe{
\item{Méthode}{Nom du modèle ("Poisson")}
\item{Ajustement (résultat du test HNP)}{Pourcentage moyen de résidus hors enveloppe}
\item{AICc}{Critère AICc}
\item{CPUE}{Valeur moyenne prédite}
\item{IC 95\%}{Intervalle de confiance autour de la prédiction}
\item{Commentaires}{Interprétation de l’ajustement}
\item{Convergence}{Booléen indiquant la convergence}
\item{Nb_iterations_HNP}{Nombre total d’itérations HNP utilisées}
}
}
\description{
Cette fonction ajuste un modèle linéaire généralisé de type Poisson (GLM)
sur les données de CPUE par station. Elle effectue également un test HNP
(Half-Normal Plot) pour évaluer la qualité de l’ajustement.
}
