% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_prepare.R
\name{cpue_prepare}
\alias{cpue_prepare}
\title{Préparer les données agrégées de CPUE par station}
\usage{
cpue_prepare(capture, specimen, group = c("tous", "femelles"))
}
\arguments{
\item{capture}{Un \code{data.frame} de captures, contenant au minimum les colonnes \code{no_station}, \code{nb_capture}, \code{nb_pese}.}

\item{specimen}{Un \code{data.frame} de spécimens, déjà filtré pour une seule espèce, les stations valides et aléatoires.}

\item{group}{Une chaîne de caractères, "tous" (valeur par défaut) ou "femelles", indiquant le groupe à analyser.}
}
\value{
Un \code{data.frame} contenant les colonnes suivantes :
\describe{
\item{no_station}{Identifiant de la station}
\item{CPUE}{Nombre de spécimens observés à la station, divisé par l’effort (station)}
\item{Group}{Libellé du groupe analysé : "Tous" ou "Femelles"}
}
}
\description{
Cette fonction prépare un tableau de CPUE (captures par unité d'effort) à l’échelle de la station,
à partir des données de captures et des spécimens observés. Elle permet de calculer soit la CPUE totale,
soit la CPUE restreinte aux femelles, selon l’argument \code{group}.
}
\details{
Elle suppose que les objets \code{capture} et \code{specimen} sont déjà filtrés pour ne contenir que les stations valides et aléatoires
(par exemple via \code{get_analysis_data()}) ainsi que les individus d’une seule espèce.
}
\examples{
capture <- tibble::tibble(
  no_station = c(1, 2),
  nb_capture = c(3, 2),
  nb_pese = c(3, 2)
)
specimen <- tibble::tibble(
  no_station = c(1, 1, 2, 2, 2),
  sexe = c("F", "M", "F", "F", "M")
)
cpue_prepare(capture, specimen, group = "femelles")

}
