% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_prepare.R
\name{cpue_prepare}
\alias{cpue_prepare}
\title{Préparer les données agrégées de CPUE par station}
\usage{
cpue_prepare(capture, specimen, group = c("tous", "femelles"))
}
\arguments{
\item{capture}{Un \code{data.frame} de captures, avec au minimum \code{no_station}, \code{nb_capture}, \code{nb_pese}.}

\item{specimen}{Un \code{data.frame} de spécimens, déjà filtré pour l’espèce et les stations valides/hasard.}

\item{group}{Une chaîne \code{"tous"} (par défaut) ou \code{"femelles"}, indiquant le groupe à analyser.}
}
\value{
Un \code{data.frame} avec les colonnes :
\describe{
\item{no_station}{Identifiant de la station}
\item{CPUE}{Nombre de spécimens observés à cette station}
\item{Group}{Libellé du groupe analysé : \code{"Tous"} ou \code{"Femelles"}}
}
}
\description{
Cette fonction prépare un tableau de CPUE par station à partir des spécimens observés
et des captures enregistrées. Elle permet de calculer soit la CPUE totale, soit la CPUE
limitée aux femelles, selon l'argument \code{group}.
}
\details{
Elle suppose que les objets \code{specimen} et \code{capture} sont déjà filtrés en amont pour ne
contenir que les stations valides et au hasard (via \code{get_analysis_data()}).
}
