% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_select_best_modele.R
\name{cpue_select_best_modele}
\alias{cpue_select_best_modele}
\title{Sélectionner le meilleur modèle CPUE selon le plus bas AICc}
\usage{
cpue_select_best_modele(tablemodele)
}
\arguments{
\item{tablemodele}{Un \code{data.frame} retourné par \code{modele_cpue_comparaison()} (format = \code{"data.frame"})}
}
\value{
Une chaîne de caractères (\code{Méthode}) correspondant au meilleur modèle
}
\description{
Cette fonction identifie automatiquement le meilleur modèle CPUE
parmi Poisson, NB1, NB2, CMP et GP, basé sur le plus bas AICc parmi
les modèles bien ajustés (ajustement HNP < 10). Si aucun modèle n’est
bien ajusté, elle sélectionne celui avec le plus bas AICc global.
}
\examples{
cpue <- cpue_prepare(capture, specimen, group = "tous")
tableau <- modele_cpue_comparaison(cpue, format = "data.frame")
cpue_select_best_modele(tableau)
}
