% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_select_best_modele.R
\name{cpue_select_best_modele}
\alias{cpue_select_best_modele}
\title{Sélectionner le meilleur modèle CPUE selon le plus bas AICc}
\usage{
cpue_select_best_modele(tablemodele)
}
\arguments{
\item{tablemodele}{Un \code{data.frame} retourné par \code{modele_cpue_comparaison(..., format = "data.frame")}.}
}
\value{
Une chaîne de caractères (\code{Méthode}) correspondant au nom du meilleur modèle sélectionné.
}
\description{
Cette fonction identifie automatiquement le meilleur modèle d’abondance (CPUE)
parmi les modèles Poisson, NB1, NB2, CMP et GP. La sélection est effectuée selon
le plus bas AICc parmi les modèles bien ajustés (ajustement HNP < 10).
Si aucun modèle ne satisfait ce critère, la fonction retourne celui avec le plus bas AICc global.
}
