% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{croissance_compare_modele}
\alias{croissance_compare_modele}
\title{Comparer trois modèles de croissance (Von Bertalanffy, Gompertz, Logistique)}
\usage{
croissance_compare_modele(data)
}
\arguments{
\item{data}{Un \code{data.frame} contenant au minimum les colonnes \code{ltm}, \code{age}
et \code{no_specimen}.}
}
\value{
Une liste contenant :
\describe{
\item{success}{Un indicateur logique précisant si l'analyse a pu être
réalisée.}
\item{data}{Un \code{data.frame} résumant les résultats des trois modèles, ou
\code{NULL} si l'analyse n'a pas pu être effectuée.}
\item{flextable}{Un objet \code{flextable} formaté pour affichage ou exportation,
ou \code{NULL} si l'analyse n'a pas pu être effectuée.}
\item{message}{Un message informatif ou explicatif, ou \code{NULL} si aucun
message n'est nécessaire.}
}
}
\description{
Cette fonction ajuste trois modèles de croissance non linéaire à un jeu de
données de spécimens et retourne un tableau comparatif des paramètres estimés,
des intervalles de confiance à 95 \% et du critère d'information corrigé
(AICc).
}
\details{
Si les données ne permettent pas la modélisation, par exemple en présence de
moins de 3 spécimens valides ou de moins de 3 âges distincts, la fonction
retourne un résultat avec \code{success = FALSE} et un message explicatif.

Si \code{vbStarts()} échoue à produire des valeurs initiales, une stratégie de
rechange est utilisée avec :
\itemize{
\item \code{Linf} = longueur du plus grand spécimen de l'échantillon
\item \code{K = 0.3}
\item \code{t0 = 0}
}

Un message est alors inclus dans la sortie afin d’informer l’utilisateur que
des valeurs initiales fixes ont été utilisées.

Si un ou plusieurs modèles ne convergent pas, la fonction retourne quand même
un tableau comparatif. Les valeurs numériques associées aux modèles non
convergents sont alors laissées à \code{NA}, les intervalles de confiance sont
laissés vides, et la colonne \code{convergence} indique \code{FALSE}.

Les modèles sont comparés uniquement à partir des ajustements convergés.
Le calcul de l'AICc est effectué seulement pour ces modèles. Si aucun modèle
ne converge, le tableau final est tout de même retourné et un message global
l'indique.

Les intervalles de confiance sont extraits individuellement pour chaque
paramètre. Lorsqu'un intervalle de confiance ne peut pas être calculé, la
valeur \code{"IC non calculable"} est utilisée.
}
