% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{croissance_compare_modele}
\alias{croissance_compare_modele}
\title{Comparer trois modèles de croissance (Von Bertalanffy, Gompertz, Logistique)}
\usage{
croissance_compare_modele(data, format = c("data.frame", "flextable"))
}
\arguments{
\item{data}{Un \code{data.frame} contenant au minimum : \code{sp}, \code{ltm}, \code{age},
\code{no_specimen}}

\item{format}{Format de sortie souhaité : \code{"data.frame"} (par défaut) ou
\code{"flextable"}}
}
\value{
Une liste contenant :
\describe{
\item{success}{Indique si l'analyse a pu être réalisée}
\item{data}{\code{data.frame} résumant les résultats des modèles}
\item{flextable}{Tableau formaté prêt pour insertion dans un document}
\item{message}{Message explicatif si l'analyse n'est pas disponible, si
des valeurs initiales de secours ont été utilisées, ou si aucun modèle n'a
convergé}
}
}
\description{
Cette fonction ajuste trois modèles de croissance non linéaire à un jeu de
données de spécimens et retourne un tableau comparatif des paramètres estimés,
des intervalles de confiance et du critère d'information corrigé (AICc).
}
\details{
Si les données ne permettent pas la modélisation (ex. moins de 3 âges
distincts), la fonction retourne un résultat avec \code{success = FALSE} et un
message explicatif.

Si \code{vbStarts()} échoue à produire des valeurs initiales, une stratégie de
rechange est utilisée avec :
\itemize{
\item \code{Linf} = longueur du plus grand spécimen de l'échantillon
\item \code{K = 0.3}
\item \code{t0 = 0}
}

Un message d’avertissement est alors retourné dans l’output afin d’informer
l’utilisateur que des valeurs initiales fixes ont été utilisées.

Si un ou plusieurs modèles ne convergent pas, la fonction retourne quand même
un tableau comparatif. Les valeurs associées aux modèles non convergents sont
alors remplacées par \code{"-"} et la colonne \code{convergence} indique explicitement
que le modèle n'a pas convergé.
}
