% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{croissance_compare_modele}
\alias{croissance_compare_modele}
\title{Comparaison de modèles de croissance (Von Bertalanffy, Gompertz, Logistique)}
\usage{
croissance_compare_modele(data, format = c("data.frame", "flextable"))
}
\arguments{
\item{data}{Un \code{data.frame} contenant les spécimens (doit inclure les colonnes \code{sp}, \code{ltm}, \code{age}, \code{no_specimen})}

\item{format}{Format de sortie : \code{"data.frame"} (par défaut) ou \code{"flextable"}}
}
\value{
Une liste contenant : \code{data} (\code{data.frame}) et \code{flextable} (\code{flextable})
}
\description{
Cette fonction ajuste trois modèles de croissance non linéaire (Von Bertalanffy, Gompertz, Logistique)
à des données de spécimens et retourne un tableau comparatif des paramètres estimés, des intervalles
de confiance et des critères de sélection de modèles (AICc).
}
\examples{
croissance_compare_modele(data = specimen)$data
croissance_compare_modele(data = specimen)$flextable
}
