% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_select_best_modele.R
\name{croissance_select_best_modele}
\alias{croissance_select_best_modele}
\title{Sélectionne le meilleur modèle de croissance basé sur le plus bas AICc}
\usage{
croissance_select_best_modele(tablemodele)
}
\arguments{
\item{tablemodele}{Un \code{data.frame} retourné par \code{croissance_compare_modele()}}
}
\value{
Une chaîne de caractères (nom du meilleur modèle) ou un message en cas d'erreur
}
\description{
Cette fonction identifie automatiquement le meilleur modèle (parmi Von Bertalanffy, Gompertz ou Logistique)
selon le critère d’AICc le plus faible.
}
