% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Enregistre un data.frame en fichier Excel (.xlsx)}
\usage{
download_data(data, path)
}
\arguments{
\item{data}{Un \code{data.frame} ou une liste de \code{data.frame}}

\item{path}{Chemin de sortie du fichier \code{.xlsx}}
}
\value{
\code{NULL} (le fichier est écrit sur disque)
}
\description{
Enregistre un data.frame en fichier Excel (.xlsx)
}
