% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissance_compare_modele.R
\name{extract_coef}
\alias{extract_coef}
\title{Extraire la valeur estimée d’un paramètre d’un modèle de croissance}
\usage{
extract_coef(mod, param_name = "Sinf")
}
\arguments{
\item{mod}{Un objet \code{nls} contenu dans \code{growth()}.}

\item{param_name}{Nom du paramètre à extraire (\code{"Sinf"}, \code{"K"} ou \code{"t0"}).}
}
\value{
Un nombre correspondant à la valeur estimée du paramètre, ou
\code{NA_real_} en cas d’erreur.
}
\description{
Fonction interne utilisée par \code{croissance_compare_modele()} pour récupérer
la valeur estimée d’un paramètre (\code{Sinf}, \code{K} ou \code{t0}) à partir d’un modèle
\code{nls} ajusté avec \code{growth()}.
}
\keyword{internal}
