% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_recapitulatif_inventaire.R
\name{generate_recapitulatif_compter_valeurs}
\alias{generate_recapitulatif_compter_valeurs}
\title{Compter les occurrences d’une valeur dans une colonne}
\usage{
generate_recapitulatif_compter_valeurs(df, col, val)
}
\arguments{
\item{df}{Un \code{data.frame} contenant les données à filtrer}

\item{col}{Le nom de la colonne (chaîne de caractères)}

\item{val}{La valeur à compter dans cette colonne}
}
\value{
Un entier représentant le nombre de lignes correspondantes
}
\description{
Fonction utilitaire interne utilisée dans le tableau récapitulatif.
}
\keyword{internal}
