% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_recapitulatif_inventaire.R
\name{generate_recapitulatif_inventaire}
\alias{generate_recapitulatif_inventaire}
\title{Générer un tableau récapitulatif d'un inventaire ichtyologique}
\usage{
generate_recapitulatif_inventaire(data_lac, data_station)
}
\arguments{
\item{data_lac}{Un \code{data.frame} contenant les métadonnées du lac (feuillet "Lac"),
déjà filtré sur un seul lac, un type de pêche et une ou plusieurs années.}

\item{data_station}{Un \code{data.frame} contenant les données des stations (feuillet "Station"),
déjà filtré de manière cohérente avec \code{data_lac}.}
}
\value{
Un \code{data.frame} structuré avec une ligne par type d'information
et une colonne par type de pêche. La première colonne est \code{"Type de pêche"}.
}
\description{
Cette fonction produit un tableau synthèse des métadonnées d'un inventaire ichtyologique
réalisé sur un lac donné, pour un type de pêche et une ou plusieurs années.
}
