% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_et_style.R
\name{geom_text_aquapop}
\alias{geom_text_aquapop}
\title{geom_text standardisé AquaPop}
\usage{
geom_text_aquapop(...)
}
\arguments{
\item{...}{Arguments passés à \code{geom_text()}}
}
\value{
Un objet ggplot layer
}
\description{
Fonction utilitaire pour afficher des étiquettes texte harmonisées avec le thème AquaPop.
}
