% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_pen.R
\name{get_info_pen}
\alias{get_info_pen}
\title{Obtenir les informations biologiques associées à une espèce ou un type de pêche}
\usage{
get_info_pen(input)
}
\value{
Une liste nommée contenant :
\describe{
\item{code_sp}{Code de l’espèce (ex: \code{"SAFO"})}
\item{nom_sp}{Nom complet de l’espèce (ex: \code{"Omble de fontaine"})}
\item{binwidth}{Largeur des classes pour histogramme}
\item{breaks}{Vecteur des bornes de classes}
\item{break_labels}{Étiquettes associées aux classes}
}

Retourne \code{NULL} si le code d’espèce est inconnu dans \code{pen_constants}.
}
\description{
Cette fonction retourne les métadonnées définies dans \code{pen_constants} pour une espèce cible.
L’entrée doit être un code d’espèce (\code{"SANA"}, \code{"SAFO"}, \code{"SAVI"}) ou un type de pêche
(\code{"PENT"}, \code{"PENOF"}, \code{"PENDJ"}), auquel cas le code d’espèce sera automatiquement déduit.
}
\examples{
get_info_pen("SAFO")
get_info_pen("PENT")

}
