% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_growth.R
\name{gompertz_function}
\alias{gompertz_function}
\title{Fonction de croissance de Gompertz}
\usage{
gompertz_function(age, linf, k, t0)
}
\arguments{
\item{age}{Âge du spécimen}

\item{linf}{Longueur asymptotique (mm)}

\item{k}{Taux de croissance}

\item{t0}{Âge à inflexion}
}
\value{
Longueur théorique à cet âge
}
\description{
Calcule la longueur en fonction de l’âge selon le modèle de Gompertz.
}
\keyword{internal}
