% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recolte.R
\name{load_recolte}
\alias{load_recolte}
\title{Charger et structurer les données du feuillet "Récolte" (version robuste)}
\usage{
load_recolte(path, namesheet = "Recolte", verbose = TRUE)
}
\arguments{
\item{path}{Chemin complet vers le fichier Excel (.xlsx) à importer.}

\item{namesheet}{Nom du feuillet contenant les données de récolte (défaut : "Recolte").}

\item{verbose}{Afficher les messages de transformation ? (défaut : TRUE)}
}
\value{
Un \code{data.frame} structuré contenant :
\describe{
\item{no_lac, typ_pech, no_station, sp}{Facteurs}
\item{annee}{Numérique (entier), année de récolte}
\item{nb_capture, nb_pese}{Numériques}
\item{comments_recolte}{Texte, si présente}
}
}
\description{
Cette fonction importe les données brutes du feuillet "Récolte" d’un fichier Excel
formaté selon les standards AquaPop. Elle applique des transformations pour nettoyer
les colonnes, corriger les années (Excel ou texte), et produire un tableau structuré.
}
\details{
Colonnes obligatoires attendues : \code{no_lac}, \code{typ_pech}, \code{no_station}, \code{sp},
\code{annee}, \code{nb_capture}, \code{nb_pese}.
Colonnes optionnelles : \code{comments}, qui sera renommée \code{comments_recolte} si présente.

Les colonnes peuvent apparaître dans n'importe quel ordre. La colonne \code{nom_lac}, si présente, est supprimée.
}
