% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_station.R
\name{load_station}
\alias{load_station}
\title{Charger et structurer les données du feuillet "Stations" (version robuste)}
\usage{
load_station(path, namesheet = "Stations", verbose = TRUE)
}
\arguments{
\item{path}{Chemin complet vers le fichier Excel (.xlsx) à importer.}

\item{namesheet}{Nom du feuillet contenant les données de stations (par défaut \code{"Stations"}).}

\item{verbose}{Afficher les messages de diagnostic ? (défaut \code{TRUE})}
}
\value{
Un \code{data.frame} structuré avec au minimum :
\describe{
\item{no_lac, typ_pech, no_station, st_valide, st_hasard, type_maill}{Facteurs}
\item{annee}{Année (entier)}
\item{lat_dd.dec, long_dd.dec, prof_deb, prof_fin}{Numériques}
\item{date_pose, date_leve, h_pose, h_leve, pose, leve, duree}{Variables temporelles}
\item{comments_station}{Chaîne de caractères}
}
}
\description{
Cette fonction importe les données brutes du feuillet "Stations" d’un fichier Excel,
applique des transformations pour nettoyer et standardiser les colonnes essentielles,
et calcule des variables supplémentaires utiles pour l’analyse (ex. date/heure de pose/levée, durée).
}
