% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_growth.R
\name{logistic_function}
\alias{logistic_function}
\title{Fonction de croissance logistique}
\usage{
logistic_function(age, linf, k, t0)
}
\arguments{
\item{age}{Âge du spécimen}

\item{linf}{Longueur asymptotique (mm)}

\item{k}{Taux de croissance}

\item{t0}{Âge à inflexion}
}
\value{
Longueur théorique à cet âge
}
\description{
Calcule la longueur en fonction de l’âge selon le modèle logistique.
}
\keyword{internal}
