% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_compare_modele.R
\name{maturite_compare_modele}
\alias{maturite_compare_modele}
\title{Évalue et sélectionne les modèles L50 (ou A50 si variable = "age")}
\usage{
maturite_compare_modele(
  specimen_data,
  prefer_combined = FALSE,
  variable = c("ltm", "age")
)
}
\arguments{
\item{specimen_data}{Données brutes (spécimens)}

\item{prefer_combined}{Logique : forcer l'utilisation des modèles combinés ?}

\item{variable}{Variable quantitative à utiliser : \code{"ltm"} (par défaut) ou \code{"age"}}
}
\value{
Une liste contenant :
\itemize{
\item \code{table} : liste avec \code{df} et \code{flextable} pour le tableau principal
\item \code{best_model} : une liste (ou une liste de deux) avec \code{modele}, \code{lien}, \code{variable}
\item \code{message} : texte explicatif
\item \code{table_sep} : liste avec \code{df} et \code{flextable} pour les modèles séparés
\item \code{table_comb} : liste avec \code{df} et \code{flextable} pour les modèles combinés
}
}
\description{
Évalue et sélectionne les modèles L50 (ou A50 si variable = "age")
}
