% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_compare_modele.R
\name{maturite_compare_modele}
\alias{maturite_compare_modele}
\title{Évaluer et comparer les modèles L50 (ou A50) de maturité}
\usage{
maturite_compare_modele(
  specimen_data,
  prefer_combined = FALSE,
  variable = c("ltm", "age")
)
}
\arguments{
\item{specimen_data}{Un \code{data.frame} contenant les données brutes de spécimens,
incluant les colonnes \code{maturite}, \code{sexe}, et la variable quantitative choisie (\code{ltm} ou \code{age}).}

\item{prefer_combined}{Logique indiquant si l’approche combinée doit être forcée (défaut : \code{FALSE})}

\item{variable}{Variable quantitative utilisée dans les modèles : \code{"ltm"} (par défaut) ou \code{"age"}}
}
\value{
Une liste contenant :
\describe{
\item{table}{Liste avec \code{df} et \code{flextable} pour le tableau principal (selon l’approche retenue)}
\item{best_model}{Liste contenant les 3 éléments : \code{best_model_M}, \code{best_model_F}, \code{best_model_combined}}
\item{message}{Texte interprétatif décrivant la sélection et les cas d’échec éventuels}
\item{table_sep}{Liste avec \code{df} et \code{flextable} pour les modèles séparés}
\item{table_comb}{Liste avec \code{df} et \code{flextable} pour les modèles combinés}
}
}
\description{
Cette fonction ajuste, évalue et sélectionne les modèles de maturité sexuelle
selon l’approche séparée (par sexe) et combinée (sexes confondus), à partir
des données de spécimens. Elle retourne les tableaux d’évaluation au format brut
et flextable, un message d’interprétation, et le meilleur modèle recommandé
selon l’approche choisie ou forcée (\code{prefer_combined}).
}
