% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_eval_modele.R
\name{maturite_eval_modele}
\alias{maturite_eval_modele}
\title{Évaluer l'ajustement des modèles de maturité (L50 ou A50)}
\usage{
maturite_eval_modele(models)
}
\arguments{
\item{models}{Liste des modèles retournée par \code{maturite_generate_modele()}}
}
\value{
Un \code{data.frame} trié par AICc avec les indicateurs : convergence, tests d’ajustement, commentaire
}
\description{
Cette fonction compile les critères d’évaluation de plusieurs modèles de maturité
(logistique, probit, etc.) pour les comparer selon leur qualité d’ajustement.
}
