% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_eval_modele.R
\name{maturite_eval_modele}
\alias{maturite_eval_modele}
\title{Évalue l'ajustement des modèles L50}
\usage{
maturite_eval_modele(models)
}
\arguments{
\item{models}{Liste des modèles retournée par fit_L50_models() ou fit_L50_combined_models()}
}
\value{
Un dataframe avec les critères de convergence et d'ajustement, trié par ordre croissant d'AICc.
}
\description{
Évalue l'ajustement des modèles L50
}
