% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_fit_separated_modele.R
\name{maturite_fit_separated_modele}
\alias{maturite_fit_separated_modele}
\title{Ajuste les modèles de maturité séparés par sexe (L50 ou A50)}
\usage{
maturite_fit_separated_modele(df, variable = c("ltm", "age"))
}
\arguments{
\item{df}{Un dataframe contenant les colonnes \code{maturite}, \code{sexe} et la variable quantitative (\code{ltm} ou \code{age})}

\item{variable}{Variable quantitative à utiliser : \code{"ltm"} (par défaut) ou \code{"age"}}
}
\value{
Une liste contenant les modèles logistiques (\code{logit}, \code{probit}, \code{cloglog}) pour M et F.
}
\description{
Ajuste les modèles de maturité séparés par sexe (L50 ou A50)
}
