% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_generate_modele.R
\name{maturite_generate_modele}
\alias{maturite_generate_modele}
\title{Ajuste un modèle de maturité (L50 ou A50)}
\usage{
maturite_generate_modele(
  data,
  variable = c("ltm", "age"),
  modele = c("TLO", "ADD", "COM", "INT"),
  lien = c("probit", "logit", "cloglog"),
  nboot = 10000
)
}
\arguments{
\item{data}{Jeu de données contenant les colonnes \code{maturite}, \code{sexe}, et \code{ltm} ou \code{age}}

\item{variable}{Variable quantitative à utiliser : \code{"ltm"} (par défaut) ou \code{"age"}}

\item{modele}{Type de modèle : \code{"TLO"}, \code{"ADD"}, \code{"COM"} ou \code{"INT"}}

\item{lien}{Lien à utiliser dans le glm : \code{"probit"} (par défaut), \code{"logit"}, \code{"cloglog"}}

\item{nboot}{Nombre de tirages Monte Carlo pour les intervalles (défaut : 10000)}
}
\value{
Une liste avec :
\describe{
\item{success}{Indique si le modèle a pu être ajusté}
\item{table_resultats}{Tableau brut des coefficients et points 50 \%}
\item{table_resultats_flextable}{Version formatée}
\item{commentaire}{Commentaire sur l'ajustement}
\item{message}{Message explicatif si l'analyse n'est pas disponible}
\item{graphique}{Ogive de maturité}
\item{donnees_ogive}{Données prédictives}
}
}
\description{
Cette fonction ajuste un modèle de maturité en fonction du type de modèle (\code{TLO}, \code{ADD}, \code{COM}, \code{INT})
et du lien (\code{logit}, \code{probit}, \code{cloglog}) spécifiés. Elle retourne un tableau de résultats,
un graphique de l'ogive de maturité, un commentaire sur l'ajustement, et les prédictions associées.
}
