% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_generate_modele.R
\name{maturite_generate_modele}
\alias{maturite_generate_modele}
\title{Ajuste un modèle de maturité selon le type spécifié (TLO, ADD, COM, INT) et le lien (logit, probit, cloglog)}
\usage{
maturite_generate_modele(
  data,
  variable = c("ltm", "age"),
  modele = c("TLO", "ADD", "COM", "INT"),
  lien = c("probit", "logit", "cloglog"),
  nboot = 10000
)
}
\arguments{
\item{data}{Jeu de données contenant les colonnes \code{maturite}, \code{sexe} et \code{ltm} ou \code{age}}

\item{variable}{Variable quantitative à utiliser : "ltm" (par défaut) ou "age"}

\item{modele}{Type de modèle : "TLO", "ADD", "COM" ou "INT"}

\item{lien}{Lien à utiliser dans le glm : "probit" (par défaut), "logit" ou "cloglog"}

\item{nboot}{Nombre de tirages Monte Carlo (par défaut: 10000)}
}
\value{
Liste avec \code{table_resultats}, \code{table_resultats_flextable}, \code{commentaire}, \code{graphique}, \code{donnees_ogive}
}
\description{
Ajuste un modèle de maturité selon le type spécifié (TLO, ADD, COM, INT) et le lien (logit, probit, cloglog)
}
