% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_generate_modele.R
\name{maturite_get_coef}
\alias{maturite_get_coef}
\title{Extraire un coefficient d’un modèle de maturité selon le sexe}
\usage{
maturite_get_coef(modele_glm, sexe = c("sexeF", "sexeM"), interaction = FALSE)
}
\arguments{
\item{modele_glm}{Un objet \code{glm}}

\item{sexe}{\code{"sexeF"} ou \code{"sexeM"} selon le coefficient à extraire}

\item{interaction}{Logique. Si \code{TRUE}, cible une interaction}
}
\value{
La valeur du coefficient ciblé
}
\description{
Fonction interne utilisée par \code{maturite_generate_modele()} pour extraire
un coefficient ciblé (par sexe ou interaction) à partir d’un modèle glm.
}
\keyword{internal}
