% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_select_best_combined_modele.R
\name{maturite_select_best_combined_modele}
\alias{maturite_select_best_combined_modele}
\title{Sélectionne le meilleur modèle parmi les modèles combinés}
\usage{
maturite_select_best_combined_modele(evaluation_df)
}
\arguments{
\item{evaluation_df}{Un data.frame retourné par \code{evaluate_L50_models()} appliqué aux modèles combinés.}
}
\value{
Une liste contenant :
\itemize{
\item \code{best_model} : identifiant du meilleur modèle (ou \code{NULL} si aucun n'est valide)
\item \code{message} : texte descriptif à afficher
}
}
\description{
À partir du tableau d’évaluation, cette fonction sélectionne le modèle combiné
ayant le plus bas AICc parmi ceux qui ont convergé et qui s'ajustent bien.
}
