% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturite_select_best_separated_modele.R
\name{maturite_select_best_separated_modele}
\alias{maturite_select_best_separated_modele}
\title{Sélectionne les meilleurs modèles L50 pour chaque sexe}
\usage{
maturite_select_best_separated_modele(evaluation_df)
}
\arguments{
\item{evaluation_df}{Un data.frame retourné par \code{evaluate_L50_models()}.
Il doit inclure les colonnes \code{modele_id}, \code{convergence}, \code{commentaire}, et \code{aicc}.}
}
\value{
Une liste contenant :
\itemize{
\item \code{best_model_M}, \code{best_model_F} : identifiants des meilleurs modèles par sexe (si applicables)
\item \code{use_combined} : booléen indiquant si une approche combinée est requise
\item \code{message} : texte décrivant la décision
}
}
\description{
À partir du tableau d’évaluation, cette fonction sélectionne les modèles L50 qui convergent
et qui ont un bon ajustement pour les mâles et les femelles, en choisissant celui avec le plus bas AICc.
Si aucun modèle n’est valide pour un des deux sexes, elle recommande de passer à l’approche combinée.
}
