% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_chaprob.R
\name{mortalite_chaprob}
\alias{mortalite_chaprob}
\title{Appliquer la méthode de Chapman-Robson pour estimer Z et A}
\usage{
mortalite_chaprob(specimen, pp, age_max)
}
\arguments{
\item{specimen}{Un \code{data.frame} produit par \code{mortalite_prepare_corr()}, contenant une colonne \code{age}.}

\item{pp}{Valeur du peak-plus.}

\item{age_max}{Âge maximum observé.}
}
\value{
Une liste avec deux éléments : \code{data} (tableau brut), \code{flextable} (tableau formaté).
}
\description{
Cette fonction applique la méthode Chapman-Robson via \code{fishmethods::agesurv()}
et retourne une liste contenant un tableau brut (\code{data.frame}) et une version formatée (\code{flextable}).
}
\examples{
res <- mortalite_chaprob(specimen, pp = 2, age_max = 8)
res$data
res$flextable
}
