% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_compare_modele.R
\name{mortalite_compare_modele}
\alias{mortalite_compare_modele}
\title{Comparer plusieurs modèles de mortalité (Z) et recommander le meilleur}
\usage{
mortalite_compare_modele(data)
}
\arguments{
\item{data}{Un \code{data.frame} contenant les colonnes \code{age} et \code{number},
tel que produit par la fonction \code{mortalite_prepare_extended()}.}
}
\value{
Une liste contenant :
\describe{
\item{data}{Le tableau comparatif des modèles sous forme de \code{data.frame}}
\item{flextable}{Le même tableau formaté avec \code{flextable} pour affichage}
}
}
\description{
Ajuste cinq modèles (Poisson, NB1, NB2, CMP, GP) sur les fréquences d’âge
et retourne un tableau comparatif avec AICc, HNP, estimation de Z et A (\%), etc.
}
