% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_fit_best_modele.R
\name{mortalite_fit_best_modele}
\alias{mortalite_fit_best_modele}
\title{Extraire le modèle de mortalité correspondant à la meilleure méthode sélectionnée}
\usage{
mortalite_fit_best_modele(data, methode = NULL)
}
\arguments{
\item{data}{Un data.frame contenant \code{age} et \code{number}(habituellement df_age_etendue)}

\item{methode}{Optionnel. Si fourni (e.g. "NB2"), retourne ce modèle directement.}
}
\value{
Un objet de classe \code{glm}, \code{glm.nb} ou \code{glmmTMB}
}
\description{
Cette fonction ajuste les 5 modèles de mortalité, sélectionne le meilleur
via \code{mortalite_select_best_modele()} et retourne l’objet \code{modele} correspondant.
}
