% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_fit_best_modele.R
\name{mortalite_fit_best_modele}
\alias{mortalite_fit_best_modele}
\title{Extraire le modèle de mortalité correspondant à la meilleure méthode sélectionnée}
\usage{
mortalite_fit_best_modele(data, methode = NULL)
}
\arguments{
\item{data}{Un \code{data.frame} contenant au minimum les colonnes \code{age} et \code{number}
(habituellement produit par \code{mortalite_prepare()})}

\item{methode}{Chaîne de caractères optionnelle ("poisson", "nb1", "nb2", "cmp", "gp").
Si précisée, ce modèle est ajusté directement sans comparaison.}
}
\value{
Un objet de classe \code{glm}, \code{glm.nb} ou \code{glmmTMB}, selon la méthode choisie
}
\description{
Cette fonction ajuste les cinq modèles de mortalité (Poisson, NB1, NB2, CMP, GP),
sélectionne automatiquement le meilleur via \code{mortalite_select_best_modele()},
et retourne l'objet \code{modele} correspondant. Si \code{methode} est précisé, il est utilisé directement.
}
