% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_fit_modele_poisson.R
\name{mortalite_fit_modele_poisson}
\alias{mortalite_fit_modele_poisson}
\title{Ajuster un modèle de mortalité de type Poisson}
\usage{
mortalite_fit_modele_poisson(df_age_etendue)
}
\arguments{
\item{df_age_etendue}{Un \code{data.frame} produit par \code{mortalite_prepare_extended()} contenant au minimum :
\describe{
\item{age}{Âge des individus}
\item{number}{Nombre d'individus observés}
}}
}
\value{
Un \code{data.frame} d'une ligne contenant :
\describe{
\item{methode}{Nom du modèle (\code{"poisson"})}
\item{ajustement_hnp}{Pourcentage moyen d'observations hors bande (test HNP)}
\item{aicc}{Critère d'information corrigé (aicc)}
\item{Z}{Coefficient d'âge estimé}
\item{SE}{Erreur standard associée à Z}
\item{A}{Taux de mortalité annuel estimé en \%}
\item{IC 95\%}{Intervalle de confiance du taux A}
\item{commentaire}{Appréciation qualitative de l'ajustement}
\item{convergence}{Convergence du modèle (toujours TRUE pour GLM Poisson)}
\item{nb_iterations_hnp}{Nombre de simulations HNP (2 ou 5)}
}
}
\description{
Cette fonction ajuste un modèle de mortalité sur les données étendues de fréquence d'âge
à l'aide d'un GLM Poisson. Elle applique aussi un test HNP (Half-Normal Plot) avec 2 à 5 simulations
selon la qualité de l'ajustement. Elle retourne un résumé des statistiques du modèle.
}
\examples{
df_fake <- tibble::tibble(age = 1:6, number = c(180, 120, 70, 40, 25, 10))
mortalite_fit_modele_poisson(df_fake)

}
