% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_phrase_resume.R
\name{mortalite_phrase_resume}
\alias{mortalite_phrase_resume}
\title{Générer une phrase descriptive pour le modèle de mortalité sélectionné}
\usage{
mortalite_phrase_resume(data_comparaison, modele_nom)
}
\arguments{
\item{data_comparaison}{data.frame contenant les colonnes \code{methode} et \code{A}}

\item{modele_nom}{Nom du modèle à décrire (ex: "NB1")}
}
\value{
Une chaîne de caractères résumant le modèle et la mortalité annuelle
}
\description{
Générer une phrase descriptive pour le modèle de mortalité sélectionné
}
\examples{
df <- data.frame(methode = c("NB1", "Poisson"), A = c(37, 51))
mortalite_phrase_resume(df, "NB1")

}
