% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_prepare_extended.R
\name{mortalite_prepare_extended}
\alias{mortalite_prepare_extended}
\title{Étendre artificiellement les données de fréquence d'âge avec des zéros}
\usage{
mortalite_prepare_extended(df_corrigee, age_max)
}
\arguments{
\item{df_corrigee}{Un \code{data.frame} avec les colonnes \code{age} et \code{number} produit par
\code{mortalite_prepare_corr()}.}

\item{age_max}{Âge maximal observé (typiquement obtenu avec \code{mortalite_get_age_max()}).}
}
\value{
Un \code{data.frame} contenant les âges observés + les âges étendus avec \code{number = 0}.
}
\description{
Cette fonction ajoute des classes d’âge fictives avec un nombre de captures nul (\code{number = 0})
au-delà de l’âge maximal observé, jusqu’à trois fois cet âge. Cette étape est décrite
dans Mainguy et Moral (2021) et permet d’améliorer l’ajustement des modèles.
}
\examples{
df_etendue <- mortalite_prepare_extended(df_corrigee, age_max = 10)
print(df_etendue)
}
