% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_prepare_extended.R
\name{mortalite_prepare_extended}
\alias{mortalite_prepare_extended}
\title{Étendre artificiellement les données de fréquence d'âge avec des zéros}
\usage{
mortalite_prepare_extended(df_corrigee, age_max)
}
\arguments{
\item{df_corrigee}{Un \code{data.frame} contenant les colonnes \code{age} (âge) et \code{number} (fréquence),
généralement produit par la fonction \code{mortalite_prepare_corr()}.}

\item{age_max}{Un entier indiquant l'âge maximal observé (souvent issu de \code{mortalite_get_age_max()}).}
}
\value{
Un \code{data.frame} combinant les âges observés avec les âges fictifs ajoutés, avec \code{number = 0}.
Les lignes sont triées par âge croissant.
}
\description{
Cette fonction ajoute des classes d'âge fictives avec un nombre de captures nul (\code{number = 0})
au-delà de l'âge maximal observé, jusqu'à trois fois cet âge. Cette étape permet d'améliorer
l'ajustement de certains modèles de mortalité (voir Mainguy & Moral, 2021).
}
\examples{
df <- data.frame(age = 2:5, number = c(4, 3, 2, 1))
mortalite_prepare_extended(df, age_max = 5)
}
