% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_select_best_modele.R
\name{mortalite_select_best_modele}
\alias{mortalite_select_best_modele}
\title{Sélectionner le meilleur modèle de mortalité selon le plus bas AICc}
\usage{
mortalite_select_best_modele(tablemodele)
}
\arguments{
\item{tablemodele}{Un \code{data.frame} retourné par \code{mortalite_compare_modele()$data}}
}
\value{
Une chaîne de caractères (\code{Méthode}) correspondant au meilleur modèle.
}
\description{
Cette fonction identifie automatiquement le meilleur modèle de mortalité
parmi Poisson, NB1, NB2, CMP et GP, basé sur le plus bas AICc parmi
les modèles bien ajustés (ajustement HNP < 10). Si aucun modèle n’est
bien ajusté, elle sélectionne celui avec le plus bas AICc global.
}
\examples{
mortalite_compare_modele_res_data <- mortalite_compare_modele(data = df_age_etendue)$data
mortalite_select_best_modele(mortalite_compare_modele_res_data)
}
