% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalite_test_surdispersion_poisson.R
\name{mortalite_test_surdispersion_poisson}
\alias{mortalite_test_surdispersion_poisson}
\title{Tester la sur-dispersion dans un modèle de Poisson (mortalité)}
\usage{
mortalite_test_surdispersion_poisson(df)
}
\arguments{
\item{df}{Un data.frame contenant les colonnes \code{age} et \code{number}}
}
\value{
Une liste avec \code{dispersion}, \code{message}, \code{plot}
}
\description{
Cette fonction applique \code{dispersiontest()} à un modèle de type Poisson
ajusté sur les fréquences d’âge (\code{number ~ age}), et retourne une liste contenant :
\itemize{
\item la valeur de dispersion,
\item un message interprétatif,
\item un graphique des résidus vs ajustés.
}
}
\examples{
mortalite_test_surdispersion_poisson(df)
}
