% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{pen_constants}
\alias{pen_constants}
\title{Constantes associees aux especes suivies par PEN}
\format{
Un \code{tibble} avec une ligne par espece.
}
\usage{
pen_constants
}
\description{
Contient le nom commun, le binwidth recommande pour les histogrammes,
les seuils de classes PSD (\code{breaks}) et leurs libelles (\code{break_labels}) par espece (\code{sp}).
}
\keyword{datasets}
