% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pen_constants}
\alias{pen_constants}
\title{Constantes associées aux espèces suivies par PEN}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 5 columns.
}
\usage{
pen_constants
}
\description{
Contient le nom commun, le binwidth recommandé pour les histogrammes,
les seuils de classes PSD (\code{breaks}) et leurs libellés (\code{break_labels}) par espèce (\code{sp}).
}
\keyword{datasets}
