% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_byclass.R
\name{psd_byclass}
\alias{psd_byclass}
\title{Calculer la fréquence par classe PSD pour une espèce donnée}
\usage{
psd_byclass(data)
}
\arguments{
\item{data}{Un \code{data.frame} filtré pour une seule espèce, contenant les colonnes \code{ltm} et \code{sp}.}
}
\value{
Une liste contenant :
\describe{
\item{success}{Indique si l'analyse a pu être produite}
\item{data}{Un \code{data.frame} résumant les fréquences par classe PSD}
\item{flextable}{Une version formatée du tableau}
\item{plot}{Un graphique en barres montrant la fréquence relative par classe}
\item{message}{Message explicatif si l'analyse n'est pas disponible}
}
}
\description{
Cette fonction retourne le nombre de poissons et leur fréquence relative (\%) pour chaque
classe de taille du PSD. Elle s'appuie sur les seuils définis dans \code{get_info_pen()} pour
l'espèce analysée. Les données doivent être filtrées pour contenir une seule espèce.
}
\details{
La fonction retourne un tableau de synthèse prêt à l'usage, un graphique et un \code{flextable}
formaté pour des rapports Word ou des interfaces Shiny.
}
\examples{
data_ex <- tibble::tibble(
  sp = rep("SANA", 100),
  ltm = sample(100:1000, 100, replace = TRUE)
)
psd_byclass(data_ex)

}
\seealso{
\link{get_info_pen}, \link{psd_classnames}
}
