% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_byclass.R
\name{psd_byclass}
\alias{psd_byclass}
\title{Calculer la fréquence par classe PSD pour une espèce donnée}
\usage{
psd_byclass(data)
}
\arguments{
\item{data}{Un \code{data.frame} contenant au moins les colonnes \code{ltm} et \code{sp}.}
}
\value{
Une liste contenant :
\describe{
\item{\code{data}}{Un \code{data.frame} résumant les fréquences par classe PSD.}
\item{\code{flextable}}{Une version formatée du tableau (pour Word, Shiny, etc.).}
\item{\code{plot}}{Un graphique en barres montrant la fréquence relative par classe.}
}
}
\description{
Cette fonction retourne le nombre de poissons et leur fréquence relative (\%) pour chaque
classe de taille du PSD. Les données doivent être filtrées pour une seule espèce.
}
