% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_q.R
\name{psd_q}
\alias{psd_q}
\title{Calculer l'indice PSD-Q global (Proportional Size Distribution – Quality)}
\usage{
psd_q(data)
}
\arguments{
\item{data}{Un \code{data.frame} contenant les données pour une seule espèce.}
}
\value{
Une liste nommée contenant :
\describe{
\item{success}{Indique si l'analyse a pu être produite.}
\item{data}{Un \code{data.frame} avec la valeur numérique de l'indice PSD-Q
et son intervalle de confiance à 95 \%.}
\item{flextable}{Une version formatée du tableau.}
\item{message}{Message explicatif si l'analyse n'est pas disponible.}
}
}
\description{
Cette fonction calcule l'indice PSD-Q pour une espèce cible, à partir des
longueurs des spécimens capturés.
}
\details{
L'indice PSD-Q correspond à la proportion d'individus situés dans les classes
de qualité (Q), définies par des seuils spécifiques à chaque espèce.
}
\examples{
set.seed(123)
data_ex <- data.frame(
  ltm = stats::rnorm(100, mean = 250, sd = 50),
  sp = "SAFO"
)
data_ex <- dplyr::filter(data_ex, ltm > 0)
psd_q(data_ex)

}
