% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render.R
\name{render_plot_ggplot}
\alias{render_plot_ggplot}
\title{Affiche un graphique ggplot dans Shiny avec largeur dynamique}
\usage{
render_plot_ggplot(
  output_id,
  plot,
  height = 500,
  res = 96,
  message_si_vide = NULL
)
}
\arguments{
\item{output_id}{Identifiant de \code{plotOutput()}}

\item{plot}{Objet ggplot ou \code{reactive()} le retournant}

\item{height}{Hauteur en pixels}

\item{res}{Résolution en DPI}

\item{message_si_vide}{Message à afficher si graphique vide ou invalide}
}
\description{
Affiche un graphique ggplot dans Shiny avec largeur dynamique
}
