% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render.R
\name{render_table_flextable}
\alias{render_table_flextable}
\title{Affiche un tableau flextable dans une application Shiny}
\usage{
render_table_flextable(output_id, flextable)
}
\arguments{
\item{output_id}{Identifiant utilisé dans \code{uiOutput()}}

\item{flextable}{Une expression, un objet \code{flextable}, ou une fonction \code{reactive()}}
}
\description{
Affiche un tableau flextable dans une application Shiny
}
