% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taille_masse_age.R
\name{taille_masse_age}
\alias{taille_masse_age}
\title{Génère un tableau morphologique (taille, masse, âge) au format brut et flextable}
\usage{
taille_masse_age(data)
}
\arguments{
\item{data}{Un data.frame contenant les colonnes \code{ltm}, \code{masse}, \code{age}, \code{sexe} et \code{maturite}.}
}
\value{
Une liste contenant deux éléments :
\describe{
\item{data}{Tableau brut (\code{data.frame}) des statistiques morphologiques}
\item{flextable}{Version mise en page du tableau avec \code{flextable}}
}
}
\description{
Cette fonction calcule des statistiques descriptives (N, moyenne, écart-type, min, max)
pour la longueur totale (ltm), la masse et l'âge des spécimens, regroupés par sexe et statut reproducteur.
Elle retourne à la fois un \code{data.frame} brut et un tableau \code{flextable} mis en page.
}
