% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_et_style.R
\name{theme_aquapop}
\alias{theme_aquapop}
\title{Thème ggplot standardisé AquaPop (style Quebec.ca)}
\usage{
theme_aquapop()
}
\value{
Un objet \code{theme} à ajouter à un graphique ggplot2
}
\description{
Applique un thème visuel conforme aux lignes directrices de présentation gouvernementale.
Utilise la police Google "Open Sans", des marges harmonisées et une palette neutre.
}
\examples{
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  theme_aquapop()
}
