% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_growth.R
\name{vb_function}
\alias{vb_function}
\title{Fonction de croissance de Von Bertalanffy}
\usage{
vb_function(age, linf, k, t0)
}
\arguments{
\item{age}{Âge du spécimen}

\item{linf}{Longueur asymptotique (mm)}

\item{k}{Taux de croissance}

\item{t0}{Âge hypothétique à longueur nulle}
}
\value{
Longueur théorique à cet âge
}
\description{
Calcule la longueur en fonction de l’âge selon le modèle de Von Bertalanffy.
}
