% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{wr_constants}
\alias{wr_constants}
\title{Constantes pour le calcul de l'indice de condition (Wr)}
\format{
Un \code{tibble} avec les colonnes :
\itemize{
\item \code{sp} : Code d'espece (SANA, SAFO, SAVI)
\item \code{species} : Nom anglais de l'espece
\item \code{min_TL} : Longueur minimale (mm)
\item \code{int} : Intercept de la regression log-log
\item \code{slope} : Pente de la regression log-log
\item \code{source} : Reference source
}
}
\usage{
wr_constants
}
\description{
Source : \code{wsVal()} pour touladi, omble de fontaine, dore jaune
}
\keyword{datasets}
